create stateful pytorch binary module breaks tests
