+++ otito/metrics/_base_metric.py
@@ -18,6 +18,14 @@ class BaseMetric(ABC):
     def _build_validator(self, package, config):
         return create_model(f"{package}:{self.__class__.__name__}Model", **config)
 
+    @staticmethod
+    def validation_handler(func):
+        @functools.wraps(func)
+        def inner(self, *args, **kwargs):
+            return func(self, **self._parse_input(*args, **kwargs))
+
+        return inner
+
     def _merge_args_kwargs(self, *args, **kwargs):
         arg_names = inspect.getfullargspec(self.callable).args
         arg_names.remove("self")
@@ -34,22 +42,3 @@ class BaseMetric(ABC):
 
     def __call__(self, *args, **kwargs):
         return self.callable(**self._parse_input(*args, **kwargs))
-
-
-class PyTorchBaseMetric(BaseMetric, ABC):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.callable = self.call_metric_function
-        self.reset()
-
-    @abstractmethod
-    def reset(self):
-        pass
-
-    @staticmethod
-    def validation_handler(func):
-        @functools.wraps(func)
-        def inner(self, *args, **kwargs):
-            return func(self, **self._parse_input(*args, **kwargs))
-
-        return inner
